sensors: move sensors with voting into a separate class
