+++ src/modules/sensors/CMakeLists.txt
@@ -38,6 +38,7 @@ px4_add_module(
 	STACK_MAIN 1300
 	COMPILE_FLAGS
 	SRCS
+		voted_sensors_update.cpp
 		rc_update.cpp
 		sensors.cpp
 		sensors_init.cpp
